nueva configuracion
