+++ package.json
@@ -18,11 +18,15 @@
   },
   "homepage": "https://github.com/jxebastian/cursoNode#readme",
   "dependencies": {
+    "bcrypt": "^3.0.5",
     "body-parser": "^1.18.3",
     "bootstrap": "^4.3.1",
     "express": "^4.16.4",
+    "express-session": "^1.15.6",
     "hbs": "^4.0.3",
     "jquery": "^3.3.1",
+    "memorystore": "^1.6.1",
+    "mongoose": "^5.4.22",
     "node-localstorage": "^1.3.1",
     "popper.js": "^1.14.7"
   }
