+++ scratch/session
@@ -1,1 +1,1 @@
-{"login":"Aspirante","aspirante":true,"coordinador":false}
+{"login":"Coordinador","aspirante":false,"coordinador":true}
