+++ src/config/config.js
@@ -1,0 +1,2 @@
+process.env.PORT = process.env.PORT || 3000;
+process.env.URLDB = 'mongodb://admin:a12345@ds013222.mlab.com:13222/curso_node'
