+++ template/views/inscribir-curso.hbs
