+++ template/views/registro.hbs
