+++ template/views/registroCurso.hbs
